skip removed lines in sector drawing
the previous code only checked if lines were removed in the drawLines function, but because of the move to layer feature of the select tool, you could get lines that were removed (by selecting them), sent to another layer, and then re-added (by deselecting them), causing both lines to exist in the sector. normally this is fine - the lines get checked and not drawn or interacted with in the drawLines function (called from the main draw function), but because i need to iterate through the lines in a sector and add them to an object based on color, if a removed line belonged to a removed layer that had a different line color that isn't in any current layer, it would completely break everything. lol.